src/useCases/authenticate.ts
import { APIError } from '@/errors/APIError'
import { UsersRepository } from '@/repositories/users-repository'
import { compare } from 'bcryptjs'
import { User } from '@prisma/client'

interface AuthenticateUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUseCaseReply {
  user: User
}

export class AuthenticateUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseReply> {
    const user = await this.usersRepository.findByEmail(email)

    if (!user) {
      throw new APIError('Invalid credentials', 401)
    }

    const doesPasswordMatches = await compare(password, user.password_hash)

    if (!doesPasswordMatches) {
      throw new APIError('Invalid credentials', 401)
    }

    return {
      user,
    }
  }
}
